+++ vgc/viewController.py
@@ -27,8 +27,13 @@ phi_final
 dist_from_center
 """
 import numpy as np
+from viewController_utils import *
 
 class ViewController():
+
+    # pylint: disable=too-many-instance-attributes
+    # 16 is needed for the ViewController class.
+
     """
     The purpose of this class is provide with the functionality needed
     in order to adjust the camera accordingly to how the drone moves in
@@ -92,6 +97,10 @@ class ViewController():
         self.dist_from_center = 0
 
     def update_fixhawk_input(self, roll, yaw, pitch, height, lon, lat):
+
+        # pylint: disable=too-many-arguments
+        # 7 arguments is needed for this function. 
+
         """
         Updates data from the auto pilot adapter.
         SETTER
@@ -110,7 +119,7 @@ class ViewController():
         Updates data from user interface
         SETTER
         """
-        if(not self.new_server_values):
+        if not self.new_server_values:
             if not lock_on:
                 if theta > 90:
                     self.theta_in = 89
@@ -126,94 +135,20 @@ class ViewController():
             self.lock_on = lock_on
             self.new_server_values = True
 
-    def rotation_matrix(self, roll, yaw, pitch):
-        """
-        Returns the rotation matrix given the roll(y-axis), yaw(z-axis)
-        and pitch(x-axis).
-
-        A positive roll indicates a clock-wise rotation as seen from
-        the negative side of the axis. This is the case for the pitch
-        as well. A positive yaw, however, indicates a clock-wise
-        rotation as seen from the positive side of the axis. One could
-        also say "as seen from above". This reflects the way the
-        FixHawk roll, yaw and pitch work. Note that for rotation
-        matrices the inverse and transponent are the same. So, if a
-        counter-clockwise rotation is needed, the transponent can be
-        used.
-        """
-        roll_rad = np.deg2rad(roll)
-        roll_matrix = np.matrix([
-                                [np.cos(roll_rad), 0, np.sin(roll_rad)],
-                                [0, 1, 0],
-                                [-np.sin(roll_rad), 0, np.cos(roll_rad)]
-                                ])
-        yaw_rad = np.deg2rad(yaw)
-        yaw_matrix = np.matrix([
-                                [np.cos(yaw_rad), np.sin(yaw_rad), 0],
-                                [-np.sin(yaw_rad), np.cos(yaw_rad), 0],
-                                [0, 0, 1]])
-        pitch_rad = np.deg2rad(pitch)
-        pitch_matrix = np.matrix([
-                                [1, 0, 0],
-                                [0, np.cos(pitch_rad), -np.sin(pitch_rad)],
-                                [0, np.sin(pitch_rad), np.cos(pitch_rad)]
-                                ])
-        return roll_matrix.dot(yaw_matrix.dot(pitch_matrix))
-
-    def earth_radius_at_lat(self, d_lat):
-        """
-        Input current latitude and return earth's approximate radius at
-        that position.
-        """
-        earth_radius_at_equator = 6378137
-        radius_difference_pole_equator = 21385
-        return earth_radius_at_equator - \
-        (d_lat/90) * radius_difference_pole_equator
-
-
-    def angular_to_spherical(self, theta, phi):
-        """
-        Translates an angular coordinate to a spherical.
-
-        The returned value is of type numpy.matrix(see documentation
-        for further information).
-        """
-        theta_rad = np.deg2rad(theta)
-        phi_rad = np.deg2rad(phi)
-        return np.matrix([
-                        [np.sin(theta_rad) * np.sin(phi_rad)],
-                        [np.sin(theta_rad)*np.cos(phi_rad)],
-                        [-np.cos(theta_rad)]
-                        ])
-
-    def spherical_to_angular(self, coord):
-        """
-        Translates a given spherical coordinate to an angular
-        coordinate.
-
-        Given a spherical coordinate on the form (x,y,z) this
-        function will return it in its angular form (theta, phi).
-        The input should be a numpy.matrix and the returned value is a
-        tuple.
-        """
-        phi = np.arctan2(coord.item(0), coord.item(1))
-        theta = np.arccos(-coord.item(2))
-        return np.rad2deg(theta), np.rad2deg(phi) % 360
-
-    def coordinate_to_point(self, coord1, coord2, z):
+    def coordinate_to_point(self, coord1, coord2, height):
         """
         Calculates where to look given two coordinates where coord1 is
         the origin.
 
-        Arguments coord1 and coord2 are tuples and z is the height from
-        where we are looking, in meters. Function returns the point in
-        its angular form(theta, phi). Used in lock-on mode.
+        Arguments coord1 and coord2 are tuples and height is the height
+        from where we are looking, in meters. Function returns the
+        point in its angular form(theta, phi). Used in lock-on mode.
         """
         coord_diff = (coord2[0] - coord1[0], coord2[1] - coord1[1])
         x_diff = np.tan(np.deg2rad(coord_diff[1])) * self.earth_radius_at_lat(coord1[1])
         y_diff = np.tan(np.deg2rad(coord_diff[0])) * self.earth_radius_at_lat(coord1[1])
         phi = np.arctan2(x_diff, y_diff)
-        theta_2 = np.arctan(np.sqrt(np.power(x_diff, 2) + np.power(y_diff, 2)) / z)
+        theta_2 = np.arctan(np.sqrt(np.power(x_diff, 2) + np.power(y_diff, 2)) / height)
         return np.rad2deg(theta_2), (np.rad2deg(phi) + 360) % 360
 
     def point_to_coordinate(self, theta, phi, height, d_coord):
@@ -284,7 +219,7 @@ class ViewController():
                 self.new_server_values = False
                 self.new_fixhawk_values = False
 
-    def get_image_point(self):
+    def get_image_point():
         """
         Returns our phi_final and dist_from_center. This can be used by
         cameraFilter to point out where we should look.
